requirements.dev.txt
pytest
coverage
pytest-cov
fastapi[all]
requests
python-dotenv
uvicorn
pydantic-settings
sqlalchemy
pymysql
ruff
pytest-asyncio
passlib[argon2]
python-jose[cryptography]
tenacity